fix: prevent crash from unknown workflows in detailed jobs view
